Cleanup: hardcode client_id to make it easier to configure the project
Any value seems to be accepted, and everything seems fine as long
as the value is always the same.
It was slightly harder to find this value compared to user_id and
cookie_xs since it is not always included in every request to FB.
This value is also only used when fetching / polling last active
times. This makes it easier for a user to configure the project.