create-visualisable-data-...: fix error when script is run for a long time
The shell script creates a temporary file to allow interruption with
Ctrl+C. Unfortunately, as this file was created in /tmp and since
nothing is written to it until Ctrl+C is pressed, this file could be
deleted by the system, causing all the data to be lost.

In order to fix this, the temporary file is simply created in the current
directory and care is added to make sure it is deleted (see cleanup).