fetch-user-infos: make it possible to fetch friends of users with -f
To remove code, the implementation of fetch-friend-list now extract
friends from:
	https://mbasic.facebook.com/profile.php?v=friends&id=userid
instead of:
	https://mbasic.facebook.com/friends/center/friends

add-user-infos: simplified since fetch-friend-list doesn't return ids
anymore, so it cannot be used with add-user-infos.

create-visualisable-data-from-*: add-user-infos is simplified and
arguments are directly passed to fetch-user-infos.

fetch-last-active-times-with-names is deleted since it can now
only be implemented like fetch-last-active-times-with-details.